refactor: move builtin component folder
