+++ packages/drip-table/src/builtin-components/component.ts
