+++ packages/drip-table/src/builtin-components/image/index.tsx
