+++ packages/drip-table/src/builtin-components/text/index.tsx
